requirements.txt
selenium
pyOfferUp
azure-cosmos
python-dotenv